+++ FindMeTutor/FindMeTutor/tutor_app/src/main/AndroidManifest.xml
@@ -10,17 +10,21 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/MyMaterialTheme">
-        <activity android:name="com.example.tutor.LoginActivity">
+        <activity android:name=".LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.example.tutor.HomeActivity" />
-        <activity android:name="com.example.tutor.RegisterActivity" />
-        <activity android:name="com.example.tutor.RequestActivity" />
-        <activity android:name="com.example.tutor.SubjectSettingsActivity" />
+        <activity android:name=".HomeActivity" />
+        <activity android:name=".RegisterActivity" />
+        <activity android:name=".RequestActivity" />
+        <activity android:name=".SubjectSettingsActivity" />
+        <activity
+            android:name=".NotificationsActivity"
+            android:label="@string/title_activity_notifications"
+            android:theme="@style/MyMaterialTheme"></activity>
     </application>
 
 </manifest>
