+++ FindMeTutor/FindMeTutor/tutor_app/src/main/java/com/example/tutor/tutor_getsubject.java
@@ -2,7 +2,6 @@ package com.example.tutor;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.os.AsyncTask;
 import android.widget.Toast;
 
@@ -49,7 +48,7 @@ public class tutor_getsubject extends AsyncTask<String, String, String> {
 
     static String out ;
 
-    public tutor_getsubject(HomeActivity par, String student_id , List<Notifications> obj){
+    public tutor_getsubject(Activity par, String student_id , List<Notifications> obj){
         parent = par;
         TutorID = student_id;
        in = obj ;
@@ -171,9 +170,11 @@ public class tutor_getsubject extends AsyncTask<String, String, String> {
                     studentName = jsObj.getString("student_fname");
                     studentSurname = jsObj.getString("student_lname");
                     description = jsObj.getString("description");
-                     Toast.makeText(parent.getApplicationContext(), studentName, Toast.LENGTH_SHORT).show();
+
                     in.add(new Notifications(subjectName,subjectCode,date,time,studentName,studentSurname,description,R.drawable.notify,parent));
                 }
+
+               // Toast.makeText(parent.getApplicationContext(), in.get(0).studentSurname, Toast.LENGTH_SHORT).show();
 
             } catch (JSONException e) {
                 e.printStackTrace();
@@ -188,8 +189,9 @@ public class tutor_getsubject extends AsyncTask<String, String, String> {
     public List getList(){return in;}
 
    public static void startActivity(Context context) {
-        context.startActivity(new Intent(context, NotificationsFragment.class).putExtra("subject", out));
+        //context.startActivity(new Intent(context, NotificationsFragment.class).putExtra("subject", out));
     }
+
 
 }
