+++ FindMeTutor/FindMeTutor/tutor_app/src/main/res/values/strings.xml
@@ -63,5 +63,6 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="title_activity_notifications">NotificationsActivity</string>
 
 </resources>
